+++ apps/ai-dial-admin/src/components/Adapter/Modals/DuplicateAdapter.tsx
@@ -5,7 +5,7 @@ import { ButtonsI18nKey, EntityFieldsI18nKey, EntityPlaceholdersI18nKey } from '
 import { useI18n } from '@/src/locales/client';
 import { DialAdapter } from '@/src/models/dial/adapter';
 import DisplayNameControl from '@/src/components/BaseControls/DisplayName';
-import IdControl from '@/src/components/BaseControls/Id';
+import IdControl from '@/src/components/BaseControls/Id/Id';
 import EndpointControl from '@/src/components/BaseControls/Endpoint/Endpoint';
 import { useSaveValidationContext, ValidationActionType } from '@/src/context/SaveValidationContext';
 import { ApplicationRoute } from '@/src/types/routes';
