+++ apps/ai-dial-admin/src/components/EntityMainProperties/Properties/EntityProperties.tsx
@@ -4,7 +4,7 @@ import { FC, useCallback, useState } from 'react';
 import { getInterceptorTemplatesList } from '@/src/app/[lang]/interceptor-templates/actions';
 import DescriptionControl from '@/src/components/BaseControls/Description';
 import DisplayNameControl from '@/src/components/BaseControls/DisplayName';
-import IdControl from '@/src/components/BaseControls/Id';
+import IdControl from '@/src/components/BaseControls/Id/Id';
 import { getSourceItems } from '@/src/components/SourceField/constants';
 import SourceField from '@/src/components/SourceField/SourceField';
 import { EntitiesI18nKey, EntityFieldsI18nKey, EntityPlaceholdersI18nKey } from '@/src/constants/i18n';
