+++ apps/ai-dial-admin/src/components/InterceptorTemplates/Modals/Duplicate.tsx
@@ -2,7 +2,7 @@ import { DialFormPopup } from '@epam/ai-dial-ui-kit';
 import { FC, useEffect, useState } from 'react';
 
 import DisplayNameControl from '@/src/components/BaseControls/DisplayName';
-import IdControl from '@/src/components/BaseControls/Id';
+import IdControl from '@/src/components/BaseControls/Id/Id';
 import { ButtonsI18nKey } from '@/src/constants/i18n';
 import { useSaveValidationContext, ValidationActionType } from '@/src/context/SaveValidationContext';
 import { useI18n } from '@/src/locales/client';
